[FIR] Move `RedundantCallOfConversionMethod` to the default checkers list
KT-77727

+++ compiler/testData/diagnostics/testsWithStdLib/resolve/kt4711.fir.kt
@@ -7,7 +7,7 @@ fun main() {
     val startTimeNanos = System.nanoTime()
 
     // the problem sits on the next line:
-    val pi = 4.0.toDouble() * delta * (1..n).reduce(
+    val pi = 4.0 * delta * (1..n).reduce(
             {t, i ->
                 val x = (i - 0.5) * delta
                 <!RETURN_TYPE_MISMATCH!>t + 1.0 / (1.0 + x * x)<!>
